olga-project/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.48
asyncpg==0.31.0
python-dotenv==1.0.1
httpx==0.27.2
